requirements.txt
torch
torchvision
facenet-pytorch
opencv-python
numpy
pillow
tqdm
pandas
requests
kaggle